sound: fix array index to type int
This changes the first argument to resampleSound(), which is used as
an index to the sounds[] array, into an int. This fixes a lot of
"array subscript has type 'char'" compiler warnings. The variables
passed over as the first argument to resampleSound() in the rest of
the code are always od type int, anyway.

While at it, remove the "unused variable 'sourceSample'" warning as
well.