setup: fix order of controls.setAxis() arguments read from config file
For whatever reason, when reading the joystick axes/direction settings
back in from the config file, the third argument to controls.setAxis()
was evaluated before the second one. This means the roles of axis and
direction were switched after a save/read config cycle.

For controls which have been disabled for joystick control, e.g.
C_PAUSE, by setting their axis to -1 this means that upon reading the
config file back in, the axis is assigned the direction value 0 and
the direction is assigned the axis value -1 converted to a boolean,
thus 1.

Axis 0 and direction 1 is, however, usually assigned to the "right"
action, which means that whenever this button was pressed the whole
bunch of other actions were triggered as well.

This fixes #24.